Buttons tweaked and Updated for creating a Setup
I added a pause/play button so that you can control the video from a different screen. The pause/play button when in full screen mode doesn't seem to work, so this is a backup.
The play all button will clear the current playlist and add the one currently in the list.
I was able to create an installer for this program at the cost of being able to debug in editor. It will now complain about a mismatch between the target platform and VLC's dll's platform. This was necessary in order to create a setup.exe using the Microsoft Visual Studio Installer Projects 2022 extension.
Tested and verified to work on both windows 11 and 10.